tasks

+++ src/main/java/com/stackroute/muzixapp/service/MuzixServiceImpl.java
@@ -1,6 +1,8 @@
 package com.stackroute.muzixapp.service;
 
 import com.stackroute.muzixapp.domain.Track;
+import com.stackroute.muzixapp.exceptions.TrackAlreadyExists;
+import com.stackroute.muzixapp.exceptions.TrackNotFound;
 import com.stackroute.muzixapp.repository.MuzixRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -18,8 +20,14 @@ public class MuzixServiceImpl implements MuzixService {
 
 
     @Override
-    public Track saveTrack(Track track) {
+    public Track saveTrack(Track track) throws TrackAlreadyExists {
+        if(muzixRepository.existsById(track.getTrackId())){
+            throw new TrackAlreadyExists("Track already exists");
+        }
         Track saveTrack = muzixRepository.save(track);
+        if(saveTrack == null){
+            throw new TrackAlreadyExists("Track already exists");
+        }
         return saveTrack;
     }
 
@@ -29,13 +37,27 @@ public class MuzixServiceImpl implements MuzixService {
     }
 
     @Override
-    public boolean updateTrack(int trackId,Track track) {
+    public boolean updateTrack(int trackId,Track track) throws TrackNotFound {
+        if(track==null){
+            throw new TrackNotFound("Track Not Found");
+        }
         track.setTrackId(trackId);
             return true;
     }
     @Override
-    public boolean removeTrack(int trackId) {
+    public boolean removeTrack(int trackId) throws TrackNotFound{
+        if(muzixRepository.existsById(trackId)==false){
+            throw new TrackNotFound("Track not found");
+        }
         muzixRepository.deleteById(trackId);
         return true;
     }
+
+    @Override
+    public List<Track> trackByName(String name) {
+        return muzixRepository.findAllTracksByName(name);
+    }
+
+
+
 }
